tests: fix subtle flakiness in multi process usdt test
The way the test runner determines that the BEFORE clause has started is
by splitting the string by whitespace and calling `pidof` with the last
substring. In this test, we launch the same program twice, so
there exists a time between the two invocations where `pidof` will
return the pid for the first process but we wouldn't have launched the
second one yet. bpftrace would then launch and only attach to the first
process and the test will time out.

The fix is somewhat hacky - we build a small C binary which runs one
instance as a child and one instance as the parent process.
(I tried bash-based fixes in previous iterations to no avail)

While we're doing this, we also fix issues with exception handling in
the test runner.